requirements.txt
click
keyring
passlib
pbr
python-ldap
pyyaml
sshpubkeys
